[#9079] Separate component tests and E2E tests as clear as possible (#9085)
* Remove Browser and BrowserPool in test.driver package

Use the copy defined in e2e package

* Separate properties used in component tests to ComponentTestProperties

Rationale: so that component tests and E2E tests use independent set of properties

* Move utility classes for E2E tests to e2e package

* Rename ComponentTestProperties back to TestProperties

The conflict with the TestProperties used for E2E tests is now no more.

* Move TestNG XML and properties files to their respective scopes

* Move base test classes used in E2E to e2e package

* Remove AllJsTests.java

We are now using Karma/Jasmine to unit test front-end code.

* jacocoReport: Only include production code

Test codes are not meaningful to be counted in coverage

* Modify gitattributes to account for e2e folder

* Update testing instructions